+++ packages/node/CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support for KYVE integration with supporting flags (#235)
-  - `--kyveEndpoint`
-  - `--kyveChainId`
-  - `--storageUrl`
+  - `--kyve-endpoint`
+  - `--kyve-chain-id`
+  - `--kyve-storage-url`
 
 ## [3.10.0] - 2024-04-10
 ### Changed