+++ packages/node/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Unused @apollo/client dependency causing version conflicts (#253)
 
+### Changed
+- Update dependencies and apply changes to match (#254)
+
 ## [3.10.0] - 2024-04-10
 ### Changed
 - Updated with node-core. Now dictionary supports multiple endpoints, indexer will fetch and switch dictionaries based on available blocks