+++ packages/node/CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `--kyve-endpoint` (default value is `https://api-us-1.kyve.network`, To disable use `false`)
   - `--kyve-chain-id` (default value is `kyve-1`)
   - `--kyve-storage-url` (default value is `https://arweave.net`) 
+### Removed
+- Unused @apollo/client dependency causing version conflicts (#253)
+### Changed
+- Update dependencies and apply changes to match (#254)
 
 ## [3.10.0] - 2024-04-10
 ### Changed